Refactor succeed and revert checks
